split RtpParameters.ts into rtpParametersTypes.ts and rtpParametersFbsUtils.ts and same for SCTP and SRTP files
